requirements.txt
gspread>=6.0
google-auth>=2.0
google-api-python-client>=2.0
openpyxl>=3.1
xlrd>=2.0
pdfplumber>=0.10
python-dotenv>=1.0
ibflex>=0.15
